# Fix #101: Remove SimpleType