--- a/internal/zinc-benchmarks/src/main/scala/xsbt/ScalacBenchmark.scala
+++ b/internal/zinc-benchmarks/src/main/scala/xsbt/ScalacBenchmark.scala
@@ -1,18 +1,18 @@
-/*package xsbt
+package xsbt
 
 import java.util.concurrent.TimeUnit
 import org.openjdk.jmh.annotations._
 
-class ShapelessBenchmark extends BenchmarkBase {
-  _project = BenchmarkProjects.Shapeless
+class ScalacBenchmark extends BenchmarkBase {
+  _project = BenchmarkProjects.Scalac
 }
 
 @State(Scope.Benchmark)
 @BenchmarkMode(Array(Mode.SingleShotTime))
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @Fork(value = 10, jvmArgs = Array("-XX:CICompilerCount=2"))
-class ColdShapelessBenchmark extends ShapelessBenchmark {
-  _subprojectToRun = _project.subprojects.head // CoreJVM
+class ColdScalacBenchmark extends ScalacBenchmark {
+  _subprojectToRun = _project.subprojects.head
   @Benchmark
   override def compile(): Unit = {
     super.compile()
@@ -25,8 +25,8 @@ class ColdShapelessBenchmark extends ShapelessBenchmark {
 @Warmup(iterations = 0)
 @Measurement(iterations = 1, time = 30, timeUnit = TimeUnit.SECONDS)
 @Fork(value = 3)
-class WarmShapelessBenchmark extends ShapelessBenchmark {
-  _subprojectToRun = _project.subprojects.head // CoreJVM
+class WarmScalacBenchmark extends ScalacBenchmark {
+  _subprojectToRun = _project.subprojects.head
   @Benchmark
   override def compile(): Unit = {
     super.compile()
@@ -39,10 +39,10 @@ class WarmShapelessBenchmark extends ShapelessBenchmark {
 @Warmup(iterations = 6, time = 10, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 6, time = 10, timeUnit = TimeUnit.SECONDS)
 @Fork(value = 3)
-class HotShapelessBenchmark extends ShapelessBenchmark {
-  _subprojectToRun = _project.subprojects.head // CoreJVM
+class HotScalacBenchmark extends ScalacBenchmark {
+  _subprojectToRun = _project.subprojects.head
   @Benchmark
   override def compile(): Unit = {
     super.compile()
   }
-}*/
+}