--- a/zinc/src/sbt-test/source-dependencies/struct-projection/B.scala
+++ b/zinc/src/sbt-test/source-dependencies/struct-projection/B.scala
@@ -1,0 +1,3 @@
+object B {
+  val x: Int = A.m
+}