--- a/zinc/src/sbt-test/source-dependencies/struct-projection/changes/A.scala
+++ b/zinc/src/sbt-test/source-dependencies/struct-projection/changes/A.scala
@@ -1,0 +1,3 @@
+object A {
+  def m: ({type T <: String})#T = ???
+}