`SimpleType` has been removed from sbt and a test has been added to show that projection on structure are correctly invalidated by sbt now.