@jvican 2017-03-08
@dwijnand This is not a breaking change because this API never got exposed to the public that I know of, it's internal to the compiler bridge.