@zehkae 2017-03-08
I made the change you asked for by amending the last commit.
I'm not yet used to GitHub's Pull Request, so please tell me if it was incorrect to use `--amend`.

I ran the benchmark on my laptop as follows:
```sh
systemctl stop ondemand.service
cpupower frequency-set --min 1.5GHz
cpupower frequency-set --max 1.5GHz
cpupower frequency-set --governor performance
sbt runBenchmarks
```
The results are [here](https://github.com/sbt/zinc/files/829242/benchmark.txt).

Also, since your bot asked me to tell: I signed the CLA :smiley: