@jvican 2017-03-08
@zehkae Thanks! Can you run it before and after the benchmark? We need to compare if something has changed and I cannot do it because the environment won't be the same :wink:.