@eed3si9n 2017-03-08
I agree that it's not a breaking change to the public Zinc API, but note that the API info does get persisted into gzip file, and is extracted by sbt to do things like test detection.