@dwijnand 2017-03-09
@jvican you're right. I've become so hyper-sensitive to breaking changes that I hadn't realised this wasn't public API.