@zehkae 2017-03-09
@jvican [Here](https://github.com/sbt/zinc/files/830614/benchmark-before.txt) are the benchmarks before the patch, and [here](https://github.com/sbt/zinc/files/830618/benchmark-after.txt) are the results after.