@jvican 2017-03-09
@zehkae This change has an unexpected change in performance. You've run the benchmarks at different times (since you've reused the benchmark of after from yesterday). I'll run this on my computer and see how it affects it.