@jvican 2017-03-10
Hey @zehkae, I'm benchmarking this PR today, will report you results soon :wink:.