@jvican 2017-03-15
I was doing benchmarks but this slipped through the cracks. I'll post results today.