@jvican 2017-03-16
Sorry for the slowliness in this PR @zehkae, but there's something that is not right. The measurements in my machine after this patch have too much fluctuation and show an increase in running time similar to the one you reported. I'll try to pin threads to get more predictable results and identify whether this decrease is real and why happens.
  
In the meanwhile, can you run benchmarks again? Use the Scala standard library and repeat the steps to set cpu frequency and governor. Make sure that you are not running any other processes while you benchmark and that they run exactly one after the other.