@jvican 2017-03-16
I have pushed two commits: one that reorders the statements (so that you don't have to do it) and the other one that adds Java Thread Affinity for the benchmarks. Please, update your master and try running the benchmarks again with JTA enabled. Make sure that JNA (Java Native Access) is installed in your computer first.
  
In the meanwhile, I'll run them, but please run them again. I'm curious to see the results.