@smarter 2017-03-16
@jvican I think you're too stringent. You shouldn't require external contributors to run benchmarks for hours on their computer without being able to do anything else. Furthermore, the benchmark results posted so far show that the difference is within the margin of error so I don't see what you hope to learn here. There are way too many factors at play here to be able to understand microscopic difference. If you want good benchmarking results, I suggest that you setup a CI that reports results for each merged PR and shows a nice graph like http://arewefastyet.com/, this will provide much better information than benchmarks run on contributors' machines.