@jvican 2017-03-16
This is a pretty intrusive change in Zinc. I think I'm not being too stringent. This software is going to be used in production real soon, so I don't want any performance regression to happen, and even less in a PR that only addresses *an issue of correctness*, it does not add anything to the current status of Zinc. I'm not happy to sacrifice 700 ms with a duplicated margin error.
  
Benchmarks in CI are even less reliable than running them in a computer, especifically if you have no control over the machines and they run in containers, which is usually the case. This is not a possibility.
  
The contributing guide clearly states that any change touching the core logic of the incremental compiler has to gone through benchmarks. For you, this may be uninteresting or unnecessary, since you develop and maintain your own bridge which is not in production. For me, as now maintainer of Zinc, it's pretty serious.
  
The fact that I want to get predictable benchmarking results and get things right should not hinder a more in-depth analysis on the consequences of this PR.
  
It's true that the author of this PR should perhaps not be involved in it. So I'll take care of this from now on. But I'm swamped in work, so this may delays some days more.