@smarter 2017-03-16
Has anyone done any profiling of zinc at all, using VisualVM or YourKit Profiler? I suspect the time you'll find much more improvement to be made from figuring out what to optimize based on profiling rather than relying on macro benchmarks results.