@jvican 2017-03-16
> Has anyone done any profiling of zinc at all, using VisualVM or YourKit Profiler? I suspect the time you'll find much more improvement to be made from figuring out what to optimize based on profiling rather than relying on macro benchmarks results.
  
Yes, I have, and it's not easy, particularly because scalac results intermingle with Zinc's. I think we need to improve our benchmarking suite and find out a way to attach Yourkit hooks in the beginning of the phases. But this requires time and it's not a priority for me now.