@romanowski 2017-03-16
@smarter my recent speedups was all based on Yourkit output. Baenchmarks are good to find any big breaks (but situation can be improved if our benchmarks produce times for zinc related phases alongside with full time).

@jvican I usually profile whole build and later on narrow down (in YourKit 'focus on method') scope to given phase.