@zehkae 2017-03-17
I ran the benchmarks again following your indications, here are the results: 
[benchmarks-before.txt](https://github.com/sbt/zinc/files/850590/benchmarks-before.txt)
[benchmarks-after.txt](https://github.com/sbt/zinc/files/850592/benchmarks-after.txt)

Sorry for not responding earlier, but I can only run the benchmarks overnight when I don't need my laptop.