@jvican 2017-04-04
I've run benchmarks on my machine because @zehkae's results were suspiciously unreasonable. I don't know how the results were showing such a negative performance impact of this change.
  
BEFORE:
```
[info] # Run complete. Total time: 00:25:57
[info] 
[info] Benchmark                                                           (_tempDir)    Mode  Cnt           Score            Error   Units
[info] HotScalacBenchmark.compile                                    /tmp/sbt_919ba19  sample   18       20721.726 ±        665.170   ms/op
[info] HotScalacBenchmark.compile:compile·p0.00                      /tmp/sbt_919ba19  sample            19528.679                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.50                      /tmp/sbt_919ba19  sample            20854.079                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.90                      /tmp/sbt_919ba19  sample            21635.898                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.95                      /tmp/sbt_919ba19  sample            21877.490                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.99                      /tmp/sbt_919ba19  sample            21877.490                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.999                     /tmp/sbt_919ba19  sample            21877.490                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.9999                    /tmp/sbt_919ba19  sample            21877.490                    ms/op
[info] HotScalacBenchmark.compile:compile·p1.00                      /tmp/sbt_919ba19  sample            21877.490                    ms/op
[info] HotScalacBenchmark.compile:·gc.alloc.rate                     /tmp/sbt_919ba19  sample   18         291.385 ±         10.283  MB/sec
[info] HotScalacBenchmark.compile:·gc.alloc.rate.norm                /tmp/sbt_919ba19  sample   18  6483756999.556 ±   38074179.544    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space            /tmp/sbt_919ba19  sample   18         288.506 ±         23.916  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm       /tmp/sbt_919ba19  sample   18  6423663957.333 ±  537833505.454    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Old_Gen               /tmp/sbt_919ba19  sample   18          13.500 ±         13.043  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Old_Gen.norm          /tmp/sbt_919ba19  sample   18   307806890.667 ±  297145775.885    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space        /tmp/sbt_919ba19  sample   18           6.972 ±          2.564  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm   /tmp/sbt_919ba19  sample   18   156339664.444 ±   59747847.126    B/op
[info] HotScalacBenchmark.compile:·gc.count                          /tmp/sbt_919ba19  sample   18         102.000                   counts
[info] HotScalacBenchmark.compile:·gc.time                           /tmp/sbt_919ba19  sample   18       22066.000                       ms
[info] WarmScalacBenchmark.compile                                   /tmp/sbt_919ba19  sample    3       53284.438 ±      39804.378   ms/op
[info] WarmScalacBenchmark.compile:compile·p0.00                     /tmp/sbt_919ba19  sample            51069.846                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.50                     /tmp/sbt_919ba19  sample            53351.547                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.90                     /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.95                     /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.99                     /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.999                    /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.9999                   /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:compile·p1.00                     /tmp/sbt_919ba19  sample            55431.922                    ms/op
[info] WarmScalacBenchmark.compile:·gc.alloc.rate                    /tmp/sbt_919ba19  sample    3         123.283 ±         99.933  MB/sec
[info] WarmScalacBenchmark.compile:·gc.alloc.rate.norm               /tmp/sbt_919ba19  sample    3  7006538144.000 ±  565872164.065    B/op
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Eden_Space           /tmp/sbt_919ba19  sample    3         116.214 ±         71.571  MB/sec
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm      /tmp/sbt_919ba19  sample    3  6608815581.333 ± 3863172275.767    B/op
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Survivor_Space       /tmp/sbt_919ba19  sample    3           2.445 ±          4.584  MB/sec
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm  /tmp/sbt_919ba19  sample    3   138730165.333 ±  166744646.238    B/op
[info] WarmScalacBenchmark.compile:·gc.count                         /tmp/sbt_919ba19  sample    3          75.000                   counts
[info] WarmScalacBenchmark.compile:·gc.time                          /tmp/sbt_919ba19  sample    3        8668.000                       ms
[info] ColdScalacBenchmark.compile                                   /tmp/sbt_919ba19      ss   10       44980.828 ±       1047.047   ms/op
[info] ColdScalacBenchmark.compile:·gc.alloc.rate                    /tmp/sbt_919ba19      ss   10         149.075 ±          3.742  MB/sec
[info] ColdScalacBenchmark.compile:·gc.alloc.rate.norm               /tmp/sbt_919ba19      ss   10  7174553936.000 ±   94085063.199    B/op
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Eden_Space           /tmp/sbt_919ba19      ss   10         140.662 ±          6.393  MB/sec
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm      /tmp/sbt_919ba19      ss   10  6770402511.200 ±  314960053.291    B/op
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Old_Gen              /tmp/sbt_919ba19      ss   10           0.002 ±          0.008  MB/sec
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Old_Gen.norm         /tmp/sbt_919ba19      ss   10      109006.400 ±     368292.066    B/op
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Survivor_Space       /tmp/sbt_919ba19      ss   10           2.996 ±          0.345  MB/sec
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm  /tmp/sbt_919ba19      ss   10   144192171.200 ±   16592896.254    B/op
[info] ColdScalacBenchmark.compile:·gc.count                         /tmp/sbt_919ba19      ss   10         252.000                   counts
[info] ColdScalacBenchmark.compile:·gc.time                          /tmp/sbt_919ba19      ss   10       29289.000                       ms
[success] Total time: 1559 s, completed Apr 4, 2017 1:41:00 PM
[success] Total time: 0 s, completed Apr 4, 2017 1:41:00 PM
```
  
AFTER:
```
[info] # Run complete. Total time: 00:25:59
[info] 
[info] Benchmark                                                            (_tempDir)    Mode  Cnt           Score            Error   Units
[info] HotScalacBenchmark.compile                                    /tmp/sbt_1eb81994  sample   18       21021.852 ±        584.833   ms/op
[info] HotScalacBenchmark.compile:compile·p0.00                      /tmp/sbt_1eb81994  sample            19964.887                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.50                      /tmp/sbt_1eb81994  sample            21189.624                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.90                      /tmp/sbt_1eb81994  sample            21917.755                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.95                      /tmp/sbt_1eb81994  sample            21978.153                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.99                      /tmp/sbt_1eb81994  sample            21978.153                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.999                     /tmp/sbt_1eb81994  sample            21978.153                    ms/op
[info] HotScalacBenchmark.compile:compile·p0.9999                    /tmp/sbt_1eb81994  sample            21978.153                    ms/op
[info] HotScalacBenchmark.compile:compile·p1.00                      /tmp/sbt_1eb81994  sample            21978.153                    ms/op
[info] HotScalacBenchmark.compile:·gc.alloc.rate                     /tmp/sbt_1eb81994  sample   18         287.482 ±          8.089  MB/sec
[info] HotScalacBenchmark.compile:·gc.alloc.rate.norm                /tmp/sbt_1eb81994  sample   18  6488495421.333 ±   37767637.669    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space            /tmp/sbt_1eb81994  sample   18         291.150 ±         22.934  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm       /tmp/sbt_1eb81994  sample   18  6570901504.000 ±  474416737.126    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Old_Gen               /tmp/sbt_1eb81994  sample   18          15.226 ±         12.359  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Old_Gen.norm          /tmp/sbt_1eb81994  sample   18   349794773.778 ±  282991988.622    B/op
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space        /tmp/sbt_1eb81994  sample   18           7.021 ±          2.916  MB/sec
[info] HotScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm   /tmp/sbt_1eb81994  sample   18   159634081.778 ±   68370763.968    B/op
[info] HotScalacBenchmark.compile:·gc.count                          /tmp/sbt_1eb81994  sample   18         106.000                   counts
[info] HotScalacBenchmark.compile:·gc.time                           /tmp/sbt_1eb81994  sample   18       23959.000                       ms
[info] WarmScalacBenchmark.compile                                   /tmp/sbt_1eb81994  sample    3       53105.481 ±       8333.743   ms/op
[info] WarmScalacBenchmark.compile:compile·p0.00                     /tmp/sbt_1eb81994  sample            52747.567                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.50                     /tmp/sbt_1eb81994  sample            52948.894                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.90                     /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.95                     /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.99                     /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.999                    /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:compile·p0.9999                   /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:compile·p1.00                     /tmp/sbt_1eb81994  sample            53619.982                    ms/op
[info] WarmScalacBenchmark.compile:·gc.alloc.rate                    /tmp/sbt_1eb81994  sample    3         123.700 ±         21.462  MB/sec
[info] WarmScalacBenchmark.compile:·gc.alloc.rate.norm               /tmp/sbt_1eb81994  sample    3  7017565357.333 ±  296877411.797    B/op
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Eden_Space           /tmp/sbt_1eb81994  sample    3         113.270 ±         69.536  MB/sec
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm      /tmp/sbt_1eb81994  sample    3  6425366538.667 ± 3358459955.066    B/op
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Survivor_Space       /tmp/sbt_1eb81994  sample    3           2.648 ±          2.103  MB/sec
[info] WarmScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm  /tmp/sbt_1eb81994  sample    3   150189434.667 ±  104657211.705    B/op
[info] WarmScalacBenchmark.compile:·gc.count                         /tmp/sbt_1eb81994  sample    3          76.000                   counts
[info] WarmScalacBenchmark.compile:·gc.time                          /tmp/sbt_1eb81994  sample    3        9262.000                       ms
[info] ColdScalacBenchmark.compile                                   /tmp/sbt_1eb81994      ss   10       44784.657 ±        967.608   ms/op
[info] ColdScalacBenchmark.compile:·gc.alloc.rate                    /tmp/sbt_1eb81994      ss   10         149.636 ±          3.113  MB/sec
[info] ColdScalacBenchmark.compile:·gc.alloc.rate.norm               /tmp/sbt_1eb81994      ss   10  7173049382.400 ±   71999309.520    B/op
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Eden_Space           /tmp/sbt_1eb81994      ss   10         139.034 ±          8.819  MB/sec
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Eden_Space.norm      /tmp/sbt_1eb81994      ss   10  6664568044.800 ±  393275917.747    B/op
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Survivor_Space       /tmp/sbt_1eb81994      ss   10           3.095 ±          0.893  MB/sec
[info] ColdScalacBenchmark.compile:·gc.churn.PS_Survivor_Space.norm  /tmp/sbt_1eb81994      ss   10   148218612.000 ±   41195067.353    B/op
[info] ColdScalacBenchmark.compile:·gc.count                         /tmp/sbt_1eb81994      ss   10         250.000                   counts
[info] ColdScalacBenchmark.compile:·gc.time                          /tmp/sbt_1eb81994      ss   10       28821.000                       ms
```
  
The results look much better! :smile:.
  
In short, this PR introduces a slowdown of 100ms while hot, while decreasing the margin error for 80ms for the average case. While warm, this PR is faster by almost 2 seconds (and margin error is dramatically decreased for the average case). Lastly, while cold this PR brings 200ms improvement. There are almost no difference in memory consumption.
  
After all, this PR does not seem to affect negatively the performance of Zinc. Given my results, I feel pretty confident merging this.
  
Thank you for all your work @zekhae. @smarter Hope this also helps you on the Dotty side.